+++ README.md
@@ -1,8 +1,8 @@
-                   
+
      __  _ __ ____  __
     |  \| |  |  \ \/ /
     |_|\__|\___//_/\_\
-                   
+
 
 An Immutable Web Framework
 
@@ -11,5 +11,5 @@ An Immutable Web Framework
 
 ###todo
 
-* [ ] create mapping from event handlers to function callbacks so entire ui tree can be stringified and saved
-* [ ] add logic to just persist todos as an alternative to saving entire ui tree
+* [/] create mapping from event handlers to function callbacks so entire ui tree can be stringified and saved
+* [/] add logic to just persist todos as an alternative to saving entire ui tree
